TEXTTAB workaround
git-svn-id: https://svn.code.sf.net/p/sauerbraten/code/src@3834 07af1644-1164-4067-b897-63bf4dc9b748

+++ engine/rendertext.cpp
@@ -147,7 +147,7 @@ float text_widthf(const char *str)
 }
 
 #define FONTTAB (4*FONTW)
-#define TEXTTAB(g) ((g) + float(FONTTAB - fmod((g), FONTTAB)))
+#define TEXTTAB(g) ((g) + float(FONTTAB - fmodf((g), FONTTAB)))
 
 void tabify(const char *str, int *numtabs)
 {
